+++ Common/src/main/java/potionstudios/byg/common/block/BYGBlockProperties.java
@@ -5,9 +5,11 @@ import net.minecraft.world.item.DyeColor;
 import net.minecraft.world.level.block.*;
 import net.minecraft.world.level.block.state.BlockBehaviour;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.material.Fluids;
 import net.minecraft.world.level.material.Material;
 import net.minecraft.world.level.material.MaterialColor;
 import net.minecraft.world.level.material.PushReaction;
+import org.jetbrains.annotations.NotNull;
 import potionstudios.byg.common.block.nether.DoubleNetherPlantBlock;
 import potionstudios.byg.common.block.nether.embur.BlueNetherrackBlock;
 import potionstudios.byg.common.block.nether.embur.EmburPlantBlock;
@@ -64,25 +66,22 @@ public class BYGBlockProperties {
         }
     }
 
-    public static class BYGEmburLily extends NetherLilyBlock {
+    public static class BYGEmburLily extends BYGLilyBlock {
         public BYGEmburLily() {
             super(BlockBehaviour.Properties.of(FIREPROOF_PLANT)
-                    .sound(SoundType.GRASS)
-                    .strength(0.0f)
-                    .noOcclusion()
+                    .sound(SoundType.LILY_PAD)
+                    .instabreak()
+                    .noOcclusion(),
+                    Fluids.LAVA
             );
-            
         }
     }
 
-    public static class BYGEnderLily extends WaterlilyBlock {
+    public static class BYGEnderLily extends BYGLilyBlock {
         public BYGEnderLily() {
-            super(BlockBehaviour.Properties.of(Material.PLANT)
-                    .sound(SoundType.LILY_PAD)
-                    .instabreak()
-                    .noOcclusion()
+            super(BlockBehaviour.Properties.copy(Blocks.LILY_PAD),
+                    Fluids.WATER
             );
-            
         }
     }
 
@@ -268,9 +267,7 @@ public class BYGBlockProperties {
             super(BlockBehaviour.Properties.of(Material.GRASS)
                     .sound(SoundType.WART_BLOCK)
                     .strength(1.0F)
-
             );
-            
         }
     }
 
@@ -597,7 +594,7 @@ public class BYGBlockProperties {
         }
 
         
-        public boolean skipRendering(BlockState state, BlockState adjacentBlockState, Direction side) {
+        public boolean skipRendering(@NotNull BlockState state, BlockState adjacentBlockState, @NotNull Direction side) {
             return adjacentBlockState.getBlock() == this || super.skipRendering(state, adjacentBlockState, side);
         }
     }
@@ -613,7 +610,7 @@ public class BYGBlockProperties {
         }
 
         
-        public boolean skipRendering(BlockState state, BlockState adjacentBlockState, Direction side) {
+        public boolean skipRendering(@NotNull BlockState state, BlockState adjacentBlockState, @NotNull Direction side) {
             return adjacentBlockState.getBlock() == this || super.skipRendering(state, adjacentBlockState, side);
         }
     }
@@ -631,7 +628,7 @@ public class BYGBlockProperties {
         }
 
         
-        public boolean skipRendering(BlockState state, BlockState adjacentBlockState, Direction side) {
+        public boolean skipRendering(@NotNull BlockState state, BlockState adjacentBlockState, @NotNull Direction side) {
             return adjacentBlockState.getBlock() == this || super.skipRendering(state, adjacentBlockState, side);
         }
     }
@@ -644,7 +641,7 @@ public class BYGBlockProperties {
                     .dynamicShape()
                     .lightLevel((state) -> 12)
             );
-            
+
         }
     }