+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Require TerraBlender 1.2.0.0 or higher.
 * Make Prairie Grass grow when bone mealing grass blocks in the Prairie Biomes. Closes #1044
 * Make Wood Fence Gates Flammable
+* Make Ender and Embur Lily's Placeable
 
 ## 1.4.7
 * Update Chinese translation.